pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "documind"
version = "0.1.0"
description = "DocuMind - AI-powered document management and RAG system"
requires-python = ">=3.10"
dependencies = [
    "openai",
    "supabase",
    "python-dotenv",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "ruff",
]

[tool.setuptools.packages.find]
where = ["src"]
